Add important ROS enviroment variables for ROS1 and ROS2 to the docker-compose file for development and deployment
Important ROS environment variables are added to the docker-compose file for development and deployment to let the user know that these variables have those default values.

Those ROS environment variables are already defined in the Dockerfile, but due to its importance it's convenient to have those explicitly listed in the docker-compose file to let the user see those.